+++ tests/DFE.FindInformationAcademiesTrusts.CypressTests/cypress/pages/trusts/trustOverviewPage.ts
@@ -2,12 +2,13 @@ class TrustOverviewPage {
 
     elements = {
         overviewHeader: () => cy.get('[data-testid="page-name"]'),
-        overviewSummaryCardContentBox: () => cy.get('[data-testid="trust-summary"]'),
+        trustSummaryCard: () => cy.get('[data-testid="trust-summary"]'),
         overviewOfstedSummaryCardContentBox: () => cy.get('[data-testid="ofsted-ratings"]'),
-        topRatingItem: () => cy.get('[data-sort-value="1"]'),
-        ratingHeader: () => cy.get('thead th').contains('Rating'),
         tableRowSortValues: () => cy.get('tbody.govuk-table__body tr td[data-sort-value]'),
         firstRowRatingText: () => cy.get('tbody.govuk-table__body tr:first-child td:first-child'),
+        detailsHeader: () => cy.get('[data-testid="page-name"]'),
+        trustDetailsCard: () => cy.get('[data-testid="trust-details-summary-card"]'),
+        referenceNumberCard: () => cy.get('[data-testid="reference-numbers-summary-card"]'),
     };
 
     public checkOverviewHeaderPresent(): this {
@@ -16,72 +17,48 @@ class TrustOverviewPage {
         return this;
     }
 
-    public checkTrustOverviewSummaryCardItemsPresent(): this {
-        this.elements.overviewSummaryCardContentBox().should('contain', 'Total academies')
-        this.elements.overviewSummaryCardContentBox().should('contain', 'Academies in each local authority')
-        this.elements.overviewSummaryCardContentBox().should('contain', 'Pupil numbers')
-        this.elements.overviewSummaryCardContentBox().should('contain', 'Pupil capacity');
+    public checkTrustSummaryCardItemsPresent(): this {
+        this.elements.trustSummaryCard().should('contain', 'Total academies')
+        this.elements.trustSummaryCard().should('contain', 'Academies in each local authority')
+        this.elements.trustSummaryCard().should('contain', 'Pupil numbers')
+        this.elements.trustSummaryCard().should('contain', 'Pupil capacity');
         return this;
     }
 
-    public checkOverviewOfstedRatingsSummaryCardItemsPresent(): this {
-        this.elements.overviewOfstedSummaryCardContentBox().should('contain', 'Outstanding')
-        this.elements.overviewOfstedSummaryCardContentBox().should('contain', 'Good')
-        this.elements.overviewOfstedSummaryCardContentBox().should('contain', 'Requires improvement')
-        this.elements.overviewOfstedSummaryCardContentBox().should('contain', 'Inadequate')
-        this.elements.overviewOfstedSummaryCardContentBox().should('contain', 'Not yet inspected');
+
+    public checkTrustSummaryCardPresent(): this {
+        this.elements.trustSummaryCard().should('be.visible');
+        this.elements.trustSummaryCard().should('contain', 'Trust summary');
         return this;
     }
 
-    public checkOfstedHeaderPresent(): this {
-        this.elements.overviewHeader().should('be.visible')
-        this.elements.overviewHeader().should('contain', 'Ofsted ratings');
+    public checkTrustDetailCardPresent(): this {
+        this.elements.trustDetailsCard().should('be.visible');
+        this.elements.trustDetailsCard().should('contain', 'Trust details');
         return this;
     }
 
-    public checkCurrentTopRatingItemIs(currentTopRating: string): this {
-        this.elements.topRatingItem().should('contain', currentTopRating);
+    public checkTrustDetailCardItemsPresent(): this {
+        this.elements.trustDetailsCard().should('contain', 'Address');
+        this.elements.trustDetailsCard().should('contain', 'Opened on');
+        this.elements.trustDetailsCard().should('contain', 'Region and territory');
+        this.elements.trustDetailsCard().should('contain', 'Information from other services');
         return this;
     }
 
-    public clickRatingHeader(): this {
-        this.elements.ratingHeader().click();
+    public checkReferenceNumbersCardPresent(): this {
+        this.elements.referenceNumberCard().should('be.visible');
+        this.elements.referenceNumberCard().should('contain', 'Reference numbers');
         return this;
     }
 
-    public checkRatingSortAscending(): this {
-        const sortValues: number[] = [];
-
-        this.elements.tableRowSortValues().each(($el) => {
-            const value = parseInt($el.attr('data-sort-value') || '', 10);
-            sortValues.push(value);
-        }).then(() => {
-            const isSortedAsc = sortValues.every((val, i, arr) => !i || val >= arr[i - 1]);
-            expect(isSortedAsc).to.be.true;
-        });
-
+    public checkReferenceNumbersCardItemsPresent(): this {
+        this.elements.referenceNumberCard().should('contain', 'UID');
+        this.elements.referenceNumberCard().should('contain', 'Group ID');
+        this.elements.referenceNumberCard().should('contain', 'UKPRN');
+        this.elements.referenceNumberCard().should('contain', 'Companies House number');
         return this;
     }
-
-    public checkRatingSortDescending(): this {
-        const sortValues: number[] = [];
-
-        this.elements.tableRowSortValues().each(($el) => {
-            const value = parseInt($el.attr('data-sort-value') || '', 10);
-            sortValues.push(value);
-        }).then(() => {
-            const isSortedDesc = sortValues.every((val, i, arr) => !i || val <= arr[i - 1]);
-            expect(isSortedDesc).to.be.true;
-        });
-
-        return this;
-    }
-
-    public checkTopResultItem(topResultText: string): this {
-        this.elements.firstRowRatingText().should('have.text', topResultText);
-        return this;
-    }
-
 
 }