+++ CHANGELOG.md
@@ -16,6 +16,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Updated wording for links on the landing page to tell users they open in new tabs
 - Updated the ofsted ratings to collect information about the ofsted subgrades
 - Remove the edit contact feature flag
+- Change Trust Navigation from side nav to top nav (service navigation component)
+- Update the Govuk frontend package to ^5.7.1
+- Moved breadcrumbs into the trust banner
+- Merged Trust details page into Trust Overview page
+- Redirect successful search to Overview page
+- Removed Ofsted summary card from Trust Overview
 
 ## [Release-11][release-11] (production-2024-10-17.3654)